requirements.txt
fastapi==0.104.1
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.1.0
Requests==2.31.0
seaborn==0.13.0
streamlit==1.26.0
